LS: Move `find_methods_for_type` from `ide` to `lang::inspect`
This is the proper location of this logic and `utils` modules should
be eradicated from existence.
