+++ crates/cairo-lang-language-server/src/ide/mod.rs
@@ -5,4 +5,3 @@ pub mod hover;
 pub mod macros;
 pub mod navigation;
 pub mod semantic_highlighting;
-pub mod utils;
