+++ crates/cairo-lang-language-server/src/lang/inspect/methods.rs
