+++ crates/cairo-lang-language-server/src/lang/inspect/mod.rs
@@ -2,3 +2,4 @@
 
 pub mod crates;
 pub mod defs;
+pub mod methods;
